requirements.txt
greenlet==1.1.2
importlib-metadata==4.8.2
python-dotenv==0.19.2
SQLAlchemy==1.4.26
typing-extensions==3.10.0.2
zipp==3.6.0
